folder restructuring. creating a types/query/mutations file for every major data object
