+++ src/schema/assets/mutations.ts
@@ -7,19 +7,6 @@ import {
   GoogleImportQueueElement,
   enqueueGooglePhotoImageBatch,
 } from "~/datasources/queues/google_import";
-import { SanityAssetRef } from "~/schema/shared/refs";
-
-builder.objectType(SanityAssetRef, {
-  description: "Representation of a Sanity Asset",
-  fields: (t) => ({
-    id: t.exposeString("id", { nullable: false }),
-    assetId: t.exposeString("assetId", { nullable: false }),
-    path: t.exposeString("path", { nullable: false }),
-    url: t.exposeString("url", { nullable: false }),
-    originalFilename: t.exposeString("originalFilename", { nullable: false }),
-    size: t.exposeInt("size", { nullable: false }),
-  }),
-});
 
 const EnqueueGoogleAlbumImportInput = builder.inputType(
   "EnqueueGoogleAlbumImportInput",
