+++ src/schema/company/mutations.ts
@@ -1,4 +1,4 @@
-import { SQL, eq, ilike } from "drizzle-orm";
+import { eq } from "drizzle-orm";
 
 import { builder } from "~/builder";
 import {
@@ -6,153 +6,9 @@ import {
   insertCompaniesSchema,
   selectCompaniesSchema,
 } from "~/datasources/db/schema";
-import { sanitizeForLikeSearch } from "~/schema/shared/helpers";
 import { CompanyRef } from "~/schema/shared/refs";
 
-const CompanyStatus = builder.enumType("CompanyStatus", {
-  values: ["active", "inactive", "draft"],
-});
-
-builder.objectType(CompanyRef, {
-  description: "Representation of a workEmail",
-  fields: (t) => ({
-    id: t.exposeString("id", { nullable: false }),
-    name: t.exposeString("name", { nullable: true }),
-    description: t.exposeString("description", { nullable: true }),
-    domain: t.exposeString("domain", { nullable: false }),
-    logo: t.exposeString("logo", { nullable: true }),
-    website: t.exposeString("website", { nullable: true }),
-    status: t.field({
-      type: CompanyStatus,
-      nullable: true,
-      description: "Not available to users",
-      // Solo superadmins pueden saber el status de una empresa
-      resolve: (root, _, { USER }) => {
-        if (USER?.isSuperAdmin && root.status) {
-          return root.status;
-        }
-        return null;
-      },
-    }),
-    hasBeenUpdated: t.field({
-      type: "Boolean",
-      nullable: false,
-      resolve: async (root, args, { DB }) => {
-        const company = await DB.query.companiesSchema.findFirst({
-          where: (c, { eq }) => eq(c.id, root.id),
-        });
-        if (!company) {
-          return false;
-        }
-        return Boolean(company.updatedAt);
-      },
-    }),
-    salarySubmissions: t.field({
-      type: "Int",
-      nullable: false,
-      resolve: async (root, args, { DB }) => {
-        const submissions = await DB.query.salariesSchema.findMany({
-          columns: {
-            id: true,
-          },
-          where: (sS, { eq }) => eq(sS.companyId, root.id),
-        });
-        return submissions.length;
-      },
-    }),
-  }),
-});
-
-const SearchCompaniesInput = builder.inputType("SearchCompaniesInput", {
-  fields: (t) => ({
-    companyName: t.field({
-      type: "String",
-      required: false,
-    }),
-    description: t.field({
-      type: "String",
-      required: false,
-    }),
-    domain: t.field({
-      type: "String",
-      required: false,
-    }),
-    website: t.field({
-      type: "String",
-      required: false,
-    }),
-  }),
-});
-
-builder.queryFields((t) => ({
-  companies: t.field({
-    description: "Get all available companies",
-    type: [CompanyRef],
-    authz: {
-      rules: ["IsAuthenticated"],
-    },
-    args: {
-      input: t.arg({
-        type: SearchCompaniesInput,
-        required: false,
-      }),
-    },
-    resolve: async (root, { input }, { DB }) => {
-      if (!input) {
-        const companies = await DB.query.companiesSchema.findMany();
-        return companies.map((c) => selectCompaniesSchema.parse(c));
-      }
-      const { companyName, description, domain, website } = input;
-      const wheres: SQL[] = [];
-      if (companyName) {
-        wheres.push(
-          ilike(companiesSchema.name, sanitizeForLikeSearch(companyName)),
-        );
-      }
-      if (description) {
-        wheres.push(
-          ilike(
-            companiesSchema.description,
-            sanitizeForLikeSearch(description),
-          ),
-        );
-      }
-      if (domain) {
-        wheres.push(
-          ilike(companiesSchema.domain, sanitizeForLikeSearch(domain)),
-        );
-      }
-      if (website) {
-        wheres.push(
-          ilike(companiesSchema.website, sanitizeForLikeSearch(website)),
-        );
-      }
-      const companies = await DB.query.companiesSchema.findMany({
-        where: (_, { and }) => and(...wheres),
-      });
-      return companies.map((c) => selectCompaniesSchema.parse(c));
-    },
-  }),
-  company: t.field({
-    description: "Get all available companies",
-    type: CompanyRef,
-    authz: {
-      rules: ["IsAuthenticated"],
-    },
-    args: {
-      companyId: t.arg({
-        type: "String",
-        required: true,
-      }),
-    },
-    resolve: async (root, { companyId }, { DB }) => {
-      const company = await DB.query.companiesSchema.findFirst({
-        where: (c, { eq }) => eq(c.id, companyId),
-      });
-      return selectCompaniesSchema.parse(company);
-    },
-  }),
-}));
+import { CompanyStatus } from "./types";
 
 const UpdateCompanyInput = builder.inputType("UpdateCompanyInput", {
   fields: (t) => ({
