+++ src/schema/index.ts
@@ -1,21 +1,49 @@
-import "./assets/assets";
-import "./allowedCurrency/allowedCurrency";
-import "./community/community";
-import "./company/company";
-import "./events/events";
-import "./eventImages/eventImages";
-import "./money/money";
-import "./salary/salary";
-import "./status/status";
-import "./tags/tags";
-import "./ticket/ticket";
-import "./user/user";
-import "./userTickets/types";
-import "./userTickets/queries";
-import "./userTickets/mutations";
-import "./workEmail/workEmail";
-import "./workRole/workRole";
-import "./workSeniority/workSeniority";
 import { builder } from "~/builder";
+import "./allowedCurrency/mutations";
+import "./allowedCurrency/queries";
+import "./allowedCurrency/types";
+import "./assets/mutations";
+import "./assets/queries";
+import "./assets/types";
+import "./community/mutations";
+import "./community/queries";
+import "./community/types";
+import "./company/mutations";
+import "./company/queries";
+import "./company/types";
+import "./eventImages/mutations";
+import "./eventImages/queries";
+import "./eventImages/types";
+import "./events/mutations";
+import "./events/queries";
+import "./events/types";
+import "./money/mutations";
+import "./money/queries";
+import "./money/types";
+import "./salary/mutations";
+import "./salary/queries";
+import "./salary/types";
+import "./status/types";
+import "./tags/mutations";
+import "./tags/queries";
+import "./tags/types";
+import "./ticket/mutations";
+import "./ticket/queries";
+import "./ticket/types";
+import "./user/mutations";
+import "./user/queries";
+import "./user/types";
+import "./userTickets/mutations";
+import "./userTickets/queries";
+import "./userTickets/types";
+import "./workEmail/mutations";
+import "./workEmail/queries";
+import "./workEmail/types";
+import "./workRole/mutations";
+import "./workRole/queries";
+import "./workRole/types";
+import "./workSeniority/mutations";
+import "./workSeniority/queries";
+import "./workSeniority/types";
 
 export const schema = builder.toSchema();
