+++ src/schema/status/types.ts
