+++ src/schema/tags/queries.ts
@@ -1,27 +1,9 @@
 import { SQL, eq, ilike } from "drizzle-orm";
-import slugify from "slugify";
 
 import { builder } from "~/builder";
 import { selectTagsSchema, tagsSchema } from "~/datasources/db/schema";
 import { sanitizeForLikeSearch } from "~/schema/shared/helpers";
 import { TagRef } from "~/schema/shared/refs";
-
-builder.objectType(TagRef, {
-  description:
-    "Representation of a tag. Tags can be associated to many things. An event, a community, etc.",
-  fields: (t) => ({
-    id: t.exposeString("id", { nullable: false }),
-    name: t.exposeString("name", { nullable: true }),
-    slug: t.field({
-      type: "String",
-      resolve: (root) =>
-        slugify(root.name.toLowerCase(), {
-          strict: true,
-        }),
-    }),
-    description: t.exposeString("description", { nullable: true }),
-  }),
-});
 
 const TagSearchInput = builder.inputType("TagSearchInput", {
   fields: (t) => ({
