+++ src/schema/user/types.ts
@@ -1,0 +1,49 @@
+import { builder } from "~/builder";
+import {
+  AllowedUserTags,
+  selectCommunitySchema,
+} from "~/datasources/db/schema";
+import { CommunityRef, UserRef } from "~/schema/shared/refs";
+
+builder.objectType(UserRef, {
+  description: "Representation of a user",
+  fields: (t) => ({
+    id: t.exposeString("id", { nullable: false }),
+    name: t.exposeString("name", { nullable: true }),
+    username: t.exposeString("username", { nullable: false }),
+    lastName: t.exposeString("lastName", { nullable: true }),
+    bio: t.exposeString("bio", { nullable: true }),
+    isSuperAdmin: t.exposeBoolean("isSuperAdmin", { nullable: true }),
+    communities: t.field({
+      type: [CommunityRef],
+      resolve: async (root, args, ctx) => {
+        const communities = await ctx.DB.query.usersSchema.findFirst({
+          where: (u, { eq }) => eq(u.id, root.id),
+          with: {
+            usersToCommunities: {
+              with: {
+                community: true,
+              },
+            },
+          },
+          orderBy(fields, operators) {
+            return operators.asc(fields.createdAt);
+          },
+        });
+        if (
+          !communities?.usersToCommunities ||
+          communities?.usersToCommunities.length === 0
+        ) {
+          return [];
+        }
+        return communities.usersToCommunities.map(({ community }) =>
+          selectCommunitySchema.parse(community),
+        );
+      },
+    }),
+  }),
+});
+
+export const SearchableUserTags = builder.enumType("SearchableUserTags", {
+  values: Object.values(AllowedUserTags),
+});
