+++ src/schema/workEmail/mutations.ts
@@ -3,7 +3,6 @@ import { eq } from "drizzle-orm";
 import { builder } from "~/builder";
 import {
   companiesSchema,
-  selectCompaniesSchema,
   confirmationTokenSchema,
   insertCompaniesSchema,
   insertConfirmationTokenSchema,
@@ -11,118 +10,8 @@ import {
   selectWorkEmailSchema,
   workEmailSchema,
 } from "~/datasources/db/schema";
-import { statusEnumOptions } from "~/datasources/db/shared";
 import { enqueueEmail } from "~/datasources/queues/mail";
-import {
-  WorkEmailRef,
-  ValidatedWorkEmailRef,
-  CompanyRef,
-} from "~/schema/shared/refs";
-
-const EmailStatusEnum = builder.enumType("EmailStatus", {
-  values: statusEnumOptions,
-});
-
-builder.objectType(WorkEmailRef, {
-  description: "Representation of a (yet to validate) work email",
-  fields: (t) => ({
-    id: t.exposeString("id", { nullable: false }),
-    isValidated: t.field({
-      type: "Boolean",
-      nullable: false,
-      resolve: (root) => root.status === "confirmed",
-    }),
-  }),
-});
-builder.objectType(ValidatedWorkEmailRef, {
-  description: "Representation of a work email associated to the current user",
-  fields: (t) => ({
-    // ID and isValidated are the same from WorkEmailRef.
-    id: t.exposeString("id", { nullable: false }),
-    isValidated: t.field({
-      type: "Boolean",
-      nullable: false,
-      resolve: (root) => root.status === "confirmed",
-    }),
-    workEmail: t.exposeString("workEmail", { nullable: false }),
-    status: t.field({
-      type: EmailStatusEnum,
-      nullable: false,
-      resolve: (root) => root.status || "pending",
-    }),
-    confirmationDate: t.field({
-      type: "DateTime",
-      nullable: true,
-      resolve: (root) =>
-        root.confirmationDate ? new Date(root.confirmationDate) : null,
-    }),
-    company: t.field({
-      type: CompanyRef,
-      nullable: true,
-      resolve: async (root, args, { DB }) => {
-        const { companyId } = root;
-        if (!companyId) {
-          return null;
-        }
-        const company = await DB.query.companiesSchema.findFirst({
-          where: (c, { eq }) => eq(c.id, companyId),
-        });
-        if (!company) {
-          return null;
-        }
-        return selectCompaniesSchema.parse(company);
-      },
-    }),
-  }),
-});
-
-builder.queryFields((t) => ({
-  workEmails: t.field({
-    description: "Get a list of validated work emails for the user",
-    type: [ValidatedWorkEmailRef],
-    authz: {
-      rules: ["IsAuthenticated"],
-    },
-    resolve: async (root, _, { DB, USER }) => {
-      if (!USER) {
-        throw new Error("No user present");
-      }
-      const workEmail = await DB.query.workEmailSchema.findMany({
-        where: (wes, { eq }) => eq(wes.userId, USER.id),
-      });
-      return workEmail.map((we) => selectWorkEmailSchema.parse(we));
-    },
-  }),
-  workEmail: t.field({
-    description: "Get a workEmail and check if its validated for this user",
-    type: WorkEmailRef,
-    authz: {
-      rules: ["IsAuthenticated"],
-    },
-    args: {
-      email: t.arg.string({ required: true }),
-    },
-    resolve: async (root, { email }, { DB, USER }) => {
-      if (!USER) {
-        throw new Error("No user present");
-      }
-      const userId = USER.id;
-      const workEmail = await DB.query.workEmailSchema.findFirst({
-        where: (wes, { and, ilike }) =>
-          and(
-            ilike(wes.workEmail, email.toLowerCase()),
-            eq(wes.userId, userId),
-          ),
-      });
-
-      if (!workEmail) {
-        throw new Error("You don't have access");
-      }
-      return selectWorkEmailSchema.parse(workEmail);
-    },
-  }),
-}));
-
+import { WorkEmailRef } from "~/schema/shared/refs";
 builder.mutationFields((t) => ({
   startWorkEmailValidation: t.field({
     description:
