+++ src/schema/workSeniority/types.ts
